JsonSerializer added; PersistentData refactoring
